Enabled roles_remap for cross environment evaluation
Cross environment roles_remap let's us train a model on one environment and then be able to load it on a different environment that has different role names. This is only for evaluation purposes. The trained agent role is overwritten during initialization of the evaluation run with an additional parameter at the Environment level called `roles_remap` that maps old_role name to new_role name